python/add-img-dimensions-requirements.txt
markdown-it-py>=3.0.0
Pillow>=10.0.0
